Refactor famiolyData into utility modules
